requirements.txt
Django
mysqlclient
django-rest-framework
dotenv